Drawing of the spike shapes still not working

+++ fileReader.h
@@ -20,6 +20,6 @@
 #include <string.h>
 #include <math.h>
 
-int readHMMFromMatfile(const char *fname, double **spikeforms, uint32_t *nspikes, uint32_t *nchs, uint32_t *nstates, float **spikes, uint32_t **cids);
-int readHMMFromHDF5file(const char *fname, double **spikeforms, uint32_t *nspikes, uint32_t *nchs, uint32_t *nstates, float **spikes, uint32_t **cids);
+int readHMMFromMatfile(const char *fname, float **spikeforms, uint32_t *nspikes, uint32_t *nchs, uint32_t *nstates, float **spikes, uint32_t **cids);
+int readHMMFromHDF5file(const char *fname, float **spikeforms, uint32_t *nspikes, uint32_t *nchs, uint32_t *nstates, float **spikes, uint32_t **cids);
